pyproject.toml
[project]
name = "stockstock"
version = "0.1.0"
description = "AI-powered automated US stock trading system using KIS API"
requires-python = ">=3.11"
dependencies = [
    # KIS brokerage
    "python-kis>=2.1.3",

    # Configuration
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # Database
    "sqlalchemy>=2.0",

    # ML / Data
    "pandas>=2.1",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "lightgbm>=4.0",
    "ta>=0.11",

    # Scheduling
    "apscheduler>=3.10,<4",

    # Telegram
    "python-telegram-bot>=21.0",

    # Utilities
    "exchange-calendars>=4.5",
    "pytz>=2024.1",

    # Logging
    "structlog>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "jupyter>=1.0",
    "matplotlib>=3.8",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"src/stockstock/strategy/model.py" = ["N806"]  # ML 관례: X, X_train 등 대문자 변수명 허용
"src/stockstock/strategy/backtest.py" = ["N806"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
